requirements.txt
streamlit
pandas
python-dateutil
jquants-api-client
openpyxl
reportlab